+++ Changelog.md
@@ -20,9 +20,12 @@ v2.40.0 (not published yet)
 - UI: Fix path handling when passed to StaxRip via command line ([#1375](/../../issues/1375), [#1377](/../../issues/1377), [#1378](/../../issues/1378))
 - UI: Fix blocking script error messages during job proceeding ([#1231](/../../issues/1231))
 - UI: Add question when adding an existing and active job with similar project file path
+- Crop: Move AutoCrop settings from Settings to Project Options
+- Crop: Add new "Time Interval" mode for AutoCrop
+    - Analyzes a frame every x seconds, which makes it more useful throughout short and long videos
 - SvtAv1EncApp: Add long path support ([#1374](/../../issues/1374))
 - Update tools
-    - ...
+    - AutoCrop v2.4
 - Update AviSynth+ plugins
     - ...
 - Update Dual plugins